feat: Implement a pause feature
Implement a pause feature, such that the user can simply press “P” (or some other key) and pause the state of the game. When pause the game, a simple sound effect should play. At the same time this sound effect plays, the music should pause, and once the user presses P again, the gameplay and the music should resume just as they were! To cap it off, display a pause icon in the middle of the screen, nice and large, so as to make it clear the game is paused.